requirements.txt
flask>=2.3
mysql-connector-python>=8.0
requests>=2.31
python-dotenv>=1.0
